changes how `install_prerequisites.py` is tested
GitHub only hosts runners for Ubuntu, Windows, and macOS. This means
that the array of runners added in c5b510f isn't actually being used.
To get around this, we use a Docker image to see if everything worked
okay.

With this commit, we also get support for testing macOS in CI (yay!).